Switch `$oldFile` and `$newFile` when computing diff of submission source files
Currently, when uploading a new submission, the "diff to previous
submission" shows the diff the wrong way around: Lines from the new
submission start with a "-" and deleted lines start with a "+". This
drives me crazy. I think I found the correct place to change the order
of the files to be diffed, but I haven't tested.